audit(erc20): inspect msg_server.go
