+++ x/erc20/keeper/mint.go
@@ -20,33 +20,46 @@ func (k Keeper) MintingEnabled(
 ) (types.TokenPair, error) {
 	params := k.GetParams(ctx)
 	if !params.EnableErc20 {
-		return types.TokenPair{}, sdkerrors.Wrap(types.ErrERC20Disabled, "module is currently disabled by governance")
+		return types.TokenPair{}, sdkerrors.Wrap(
+			types.ErrERC20Disabled, "module is currently disabled by governance",
+		)
 	}
 
 	id := k.GetTokenPairID(ctx, token)
 	if len(id) == 0 {
-		return types.TokenPair{}, sdkerrors.Wrapf(types.ErrTokenPairNotFound, "token '%s' not registered by id", token)
+		return types.TokenPair{}, sdkerrors.Wrapf(
+			types.ErrTokenPairNotFound, "token '%s' not registered by id", token,
+		)
 	}
 
 	pair, found := k.GetTokenPair(ctx, id)
 	if !found {
-		return types.TokenPair{}, sdkerrors.Wrapf(types.ErrTokenPairNotFound, "token '%s' not registered", token)
+		return types.TokenPair{}, sdkerrors.Wrapf(
+			types.ErrTokenPairNotFound, "token '%s' not registered", token,
+		)
 	}
 
 	if !pair.Enabled {
-		return types.TokenPair{}, sdkerrors.Wrapf(types.ErrERC20Disabled, "minting token '%s' is not enabled by governance", token)
+		return types.TokenPair{}, sdkerrors.Wrapf(
+			types.ErrERC20Disabled, "minting token '%s' is not enabled by governance", token,
+		)
 	}
 
 	if k.bankKeeper.BlockedAddr(receiver.Bytes()) {
-		return types.TokenPair{}, sdkerrors.Wrapf(sdkerrors.ErrUnauthorized, "%s is not allowed to receive transactions", receiver)
+		return types.TokenPair{}, sdkerrors.Wrapf(
+			sdkerrors.ErrUnauthorized, "%s is not allowed to receive transactions", receiver,
+		)
 	}
 
 	// NOTE: ignore amount as only denom is checked on IsSendEnabledCoin
 	coin := sdk.Coin{Denom: pair.Denom}
 
-	// check if minting to a recipient address other than the sender is enabled for for the given coin denom
+	// check if minting to a recipient address other than the sender is enabled
+	// for for the given coin denom
 	if !sender.Equals(receiver) && !k.bankKeeper.IsSendEnabledCoin(ctx, coin) {
-		return types.TokenPair{}, sdkerrors.Wrapf(banktypes.ErrSendDisabled, "minting '%s' coins to an external address is currently disabled", token)
+		return types.TokenPair{}, sdkerrors.Wrapf(
+			banktypes.ErrSendDisabled, "minting '%s' coins to an external address is currently disabled", token,
+		)
 	}
 
 	return pair, nil
